refactor: remove game-modes folder
